refactor: extract VerifyMarkdownTable.ts
